pyproject.toml
[tool.poetry]
name = "random-project"
version = "0.1.0"
description = ""
authors = ["Gabriel Spacki <ifprgabriel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-allauth = "^0.63.3"
dj-rest-auth = "^6.0.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
